feat:购物车

+++ src/main/java/com/baozun/shoppingcart/service/SpuService.java
@@ -1,8 +1,7 @@
 package com.baozun.shoppingcart.service;
 
-import com.baozun.shoppingcart.controller.vo.request.PageParameterRequest;
 import com.baozun.shoppingcart.controller.vo.request.SpuRequest;
-import com.baozun.shoppingcart.dao.model.SpuCategory;
+import com.baozun.shoppingcart.dao.model.Promotion;
 import com.baozun.shoppingcart.dao.model.SpuStatusEnum;
 import com.baozun.shoppingcart.controller.vo.request.UpdateSpuRequest;
 import com.baozun.shoppingcart.dao.SpuCategoryRepository;
@@ -11,15 +10,14 @@ import com.baozun.shoppingcart.dao.SpuRepository;
 import com.baozun.shoppingcart.controller.vo.request.SpuQueryRequest;
 import com.baozun.shoppingcart.exception.AppException;
 import com.baozun.shoppingcart.exception.AppExceptionCodeMsg;
-import javax.persistence.criteria.CriteriaBuilder;
-import javax.persistence.criteria.CriteriaQuery;
+import java.util.ArrayList;
+import java.util.List;
+import javax.persistence.criteria.Join;
+import javax.persistence.criteria.JoinType;
 import javax.persistence.criteria.Path;
 import javax.persistence.criteria.Predicate;
-import javax.persistence.criteria.Root;
-import javax.transaction.Transactional;
-import javax.validation.constraints.NotNull;
-import javax.validation.constraints.PositiveOrZero;
 import lombok.RequiredArgsConstructor;
+import org.apache.commons.collections4.CollectionUtils;
 import org.apache.commons.lang3.ObjectUtils;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -36,38 +34,31 @@ public class SpuService {
 
   private final SpuCategoryRepository spuCategoryRepository;
 
-  public Page<Spu> findAll(PageParameterRequest pageParameterRequest) {
-    Pageable pageable = PageRequest.of(pageParameterRequest.getPageNumber(), pageParameterRequest.getPageSize());
-    return spuRepository.findAll(pageable);
-  }
-
-  public Page<Spu> findAllByParameter(SpuQueryRequest parameter) {
-    if (
-        ObjectUtils.allNotNull(parameter.getMinPrice(), parameter.getMaxPrice())
-            && parameter.getMinPrice() > parameter.getMaxPrice()
-    ) {
-      throw new AppException(AppExceptionCodeMsg.PRICE_RANGE_ERROR);
-    }
-    Pageable pageable = PageRequest.of(parameter.getPageNumber(), parameter.getPageSize());
-    return spuRepository.findAll(parameter, pageable);
-  }
-
   public Spu findById(Integer spuId) {
     return findSpu(spuId);
   }
 
-  @Transactional
+  public Page<Spu> findAll(SpuQueryRequest parameter) {
+    if (parameter.getMinPrice() > parameter.getMaxPrice()) {
+      throw new AppException(AppExceptionCodeMsg.PRICE_RANGE_ERROR);
+    }
+    Page<Spu> spuList = findSpuList(parameter);
+    if (CollectionUtils.isEmpty(spuList.getContent())) {
+      throw new AppException(AppExceptionCodeMsg.SPU_NOT_EXIST);
+    }
+    return spuList;
+  }
+
   public Spu saveSpu(SpuRequest spuRequest) {
     if (spuRepository.findByCode(spuRequest.getCode()) > 0) {
       throw new AppException(AppExceptionCodeMsg.SPU_IS_EXIST);
     }
-    SpuCategory spuCategory = spuCategoryRepository
+    spuCategoryRepository
         .findById(spuRequest.getCategoryId())
         .orElseThrow(() -> new AppException(AppExceptionCodeMsg.CATEGORY_NOT_EXIST));
     return spuRepository.save(spuRequest.toSpu());
   }
 
-  @Transactional
   public Spu updateSpu(UpdateSpuRequest updateSpu) {
     Spu spu = findSpu(updateSpu.getId());
     if (!ObjectUtils.isEmpty(updateSpu.getCategoryId())) {
@@ -77,23 +68,25 @@ public class SpuService {
     }
     if (
         !ObjectUtils.isEmpty(updateSpu.getPrice())
-            && !ObjectUtils.isEmpty(updateSpu.getDiscount())
-            && updateSpu.getPrice() < updateSpu.getDiscount()
+            && !ObjectUtils.isEmpty(updateSpu.getBidPrice())
+            && updateSpu.getPrice() < updateSpu.getBidPrice()
     ) {
       throw new AppException(AppExceptionCodeMsg.PRICE_SETTING_ERROR);
     }
     if (
         ObjectUtils.isEmpty(updateSpu.getPrice())
-            && !ObjectUtils.isEmpty(updateSpu.getDiscount())
-            && spu.getPrice() < updateSpu.getDiscount()
+            && !ObjectUtils.isEmpty(updateSpu.getBidPrice())
+            && spu.getPrice() < updateSpu.getBidPrice()
     ) {
       throw new AppException(AppExceptionCodeMsg.PRICE_SETTING_ERROR);
     }
-    return spuRepository.saveAndFlush(updateSpu.toSpu(spu));
+    return spuRepository.save(updateSpu.toSpu(spu));
   }
 
-  @Transactional
   public void onShelves(Integer spuId) {
+    if (spuId < 0) {
+      throw new AppException(AppExceptionCodeMsg.SPU_NOT_EXIST);
+    }
     Spu spu = findSpu(spuId);
     if (spu.getStatus().equals(SpuStatusEnum.ON_SHELVES)) {
       throw new AppException(AppExceptionCodeMsg.SPU_ALREADY_ON_SHELVES);
@@ -106,8 +99,10 @@ public class SpuService {
     }
   }
 
-  @Transactional
   public void soldOut(Integer spuId) {
+    if (spuId < 0) {
+      throw new AppException(AppExceptionCodeMsg.SPU_NOT_EXIST);
+    }
     Spu spu = findSpu(spuId);
     if (spu.getStatus().equals(SpuStatusEnum.SOLD_OUT)) {
       throw new AppException(AppExceptionCodeMsg.SPU_ALREADY_SOLD_OUT);
@@ -120,8 +115,10 @@ public class SpuService {
     }
   }
 
-  @Transactional
   public void deleteSpuById(Integer spuId) {
+    if (spuId < 0) {
+      throw new AppException(AppExceptionCodeMsg.SPU_NOT_EXIST);
+    }
     spuRepository.deleteById(spuId);
   }
 
@@ -129,4 +126,40 @@ public class SpuService {
     return spuRepository.findById(spuId).orElseThrow(() -> new AppException(AppExceptionCodeMsg.SPU_NOT_EXIST));
   }
 
+  private Page<Spu> findSpuList(SpuQueryRequest parameter) {
+    Pageable pageable = PageRequest.of(parameter.getPageNumber(), parameter.getPageSize());
+    return spuRepository.findAll((Specification<Spu>) (root, query, cb) -> {
+      List<Predicate> predicates = new ArrayList<>();
+      if (ObjectUtils.isNotEmpty(parameter.getName())) {
+        Path<Object> name = root.get("name");
+        Predicate findByName = cb.like(name.as(String.class), parameter.getName());
+        predicates.add(findByName);
+      }
+      if (ObjectUtils.isNotEmpty(parameter.getStatus())) {
+        Path<Object> status = root.get("status");
+        Predicate findByStatus = cb.equal(status, parameter.getStatus());
+        predicates.add(findByStatus);
+      }
+      if (ObjectUtils.isNotEmpty(parameter.getCategoryName())) {
+        Join<Spu, SpuRepository> join = root.join("spuCategory", JoinType.INNER);
+        Predicate findByCategoryName = cb.equal(join.get("name"), parameter.getCategoryName());
+        predicates.add(findByCategoryName);
+      }
+      if (ObjectUtils.isNotEmpty(parameter.getPromotionName())) {
+        Join<Spu, Promotion> join = root.join("promotions", JoinType.LEFT);
+        Predicate findByPromotionName = cb.equal(join.get("name"), parameter.getPromotionName());
+        predicates.add(findByPromotionName);
+      }
+      if (parameter.getMaxPrice() > 0) {
+        Path<Object> discount = root.get("discount");
+        cb.between(discount.as(Integer.class), parameter.getMinPrice(), parameter.getMaxPrice());
+      }
+      if (ObjectUtils.isNotEmpty(parameter.getCreateTime())) {
+        Predicate createTime = cb.equal(root.get("createTime"), parameter.getCreateTime());
+        predicates.add(createTime);
+      }
+      predicates.forEach(predicate -> query.where(cb.and(predicate)));
+      return query.getRestriction();
+    }, pageable);
+  }
 }
